Cleanup internal config.
Adds the ability for the keybind to be set to mouse keys(keybind can also be edited from modmenu now).

Adds an enabled by default keybind to make sorting an open inv sort the portion of the inv your mouse is hovered over when triggering(ie hovering over the player portion of the inv and pressing the keybind now sorts the players inv instead of the opened one).